+++ Frends.JSON.ConvertXMLStringToJToken/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [1.2.0] - 2026-05-07
 ### Added
-- Added optional XSD input support to validate XML and preserve schema-defined array mapping during XML to JToken conversion.
+- Added optional XSD input support to ensure consistent array/object mapping 
 
 ## [1.1.0] - 2024-08-20
 ### Updated